+++ lib/src/host_callable.dart
@@ -5,7 +5,7 @@
 import 'dart:cli';
 import 'dart:io';
 
-import 'package:sass/sass.dart' as sass;
+import 'package:sass_api/sass_api.dart' as sass;
 
 import 'dispatcher.dart';
 import 'embedded_sass.pb.dart';