+++ lib/src/importer.dart
@@ -4,7 +4,7 @@
 
 import 'dart:cli';
 
-import 'package:sass/sass.dart' as sass;
+import 'package:sass_api/sass_api.dart' as sass;
 
 import 'dispatcher.dart';
 import 'embedded_sass.pb.dart' hide SourceSpan;