+++ lib/src/logger.dart
@@ -3,7 +3,7 @@
 // https://opensource.org/licenses/MIT.
 
 import 'package:path/path.dart' as p;
-import 'package:sass/sass.dart' as sass;
+import 'package:sass_api/sass_api.dart' as sass;
 import 'package:source_span/source_span.dart';
 import 'package:stack_trace/stack_trace.dart';