+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 1.0.0-beta.11
+
+* Support version 1.0.0-beta.13 of the Sass embedded protocol:
+  * Support `Value.HwbColor`.
+  * Emit colors as `Value.HslColor` if that internal representation is
+    available.
+
 ## 1.0.0-beta.10
 
 * Support version 1.0.0-beta.12 of the Sass embedded protocol: